+++ AntiLagNext/src/AntiLagNext.App/Views/SettingsView.xaml.cs
@@ -1,8 +1,0 @@
-using System.Windows.Controls;
-
-namespace AntiLagNext.App.Views;
-
-public partial class SettingsView : UserControl
-{
-    public SettingsView() => InitializeComponent();
-}